% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MV04_LM.R
\name{asymp_MV04}
\alias{asymp_MV04}
\title{Critical values from MV04 paper, p.1831}
\usage{
asymp_MV04(alpha = c(0.01, 0.05, 0.1), M = c(1, 2, 3), type = c("no",
  "const", "trend"))
}
\description{
Critical values from MV04 paper, p.1831
}
\keyword{internal}
