% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CH03_LM.R, R/CH06_LM.R
\name{av_peri}
\alias{av_peri}
\title{Averaged (tapered) periodogram. Only for internal use.}
\usage{
av_peri(X, diff_param = 1, m_peri)

av_peri(X, diff_param = 1, m_peri)
}
\description{
Averaged (tapered) periodogram. Only for internal use.

Averaged (tapered) periodogram. Only for internal use.
}
\keyword{internal}
