% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MV04_LM.R
\name{beta_N}
\alias{beta_N}
\title{Beta estimation which is consistent in the spurious case and inconsistent under cointegration introduced by MV04.}
\usage{
beta_N(X, dx, dep, N)
}
\description{
Beta estimation which is consistent in the spurious case and inconsistent under cointegration introduced by MV04.
}
\keyword{internal}
