% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N10_LM.R
\name{crit_N10}
\alias{crit_N10}
\title{Simulated critical values for N10, memory in (0.51,1.49), dim_series<13, d1=0.1}
\usage{
crit_N10(d, nbr, alpha)
}
\description{
Simulated critical values for N10, memory in (0.51,1.49), dim_series<13, d1=0.1
}
\keyword{internal}
