% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MV04_LM.R
\name{cross_Peri}
\alias{cross_Peri}
\title{Cross periodogram for variance estimation in MV04.}
\usage{
cross_Peri(X, Y, m)
}
\description{
Cross periodogram for variance estimation in MV04.
}
\keyword{internal}
