% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ddiffw}
\alias{ddiffw}
\title{Helper function that returns AR-representation of FI(d)-process.}
\usage{
ddiffw(n, d)
}
\arguments{
\item{n}{number of coefficients to be returned}

\item{d}{memory parameter}
}
\description{
returns the first \code{n} coefficients in the AR-infinity representation
of an FI(d) process
}
