% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CH03_LM.R, R/CH06_LM.R
\name{eigen_resids}
\alias{eigen_resids}
\title{Calculates eigenvalues and eigenvectors of averaged periodogram. 
Generates cointegrating residuals with help of eigenvectors.}
\usage{
eigen_resids(X, diff_param = 1, m_peri)

eigen_resids(X, diff_param = 1, m_peri)
}
\description{
Calculates eigenvalues and eigenvectors of averaged periodogram. 
Generates cointegrating residuals with help of eigenvectors.

Calculates eigenvalues and eigenvectors of averaged periodogram. 
Generates cointegrating residuals with help of eigenvectors.
}
\keyword{internal}
