% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N10_LM.R, R/NS07_LM.R
\name{elW_wrap}
\alias{elW_wrap}
\title{Helper for memory estimation.}
\usage{
elW_wrap(X, m, mean_correct)

elW_wrap(X, m, mean_correct)
}
\description{
Helper for memory estimation.

Helper for memory estimation.
}
\keyword{internal}
