% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R08_LM.R
\name{G.hat.star.star}
\alias{G.hat.star.star}
\title{Modified complex-valued element of derivative of objective function in R08 accounting for non-equal memory.}
\usage{
G.hat.star.star(X, d, m)
}
\description{
Modified complex-valued element of derivative of objective function in R08 accounting for non-equal memory.
}
\keyword{internal}
