% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{G_hat_cpp}
\alias{G_hat_cpp}
\title{Estimation of G matrix for multivariate long memory processes.}
\usage{
G_hat_cpp(peri, Lambda_cube, d_vec, m, l, q)
}
\arguments{
\item{peri}{cube containing the periodogram of the multivariate process X.}

\item{Lambda_cube}{cube containing the Lambda matrices.}

\item{d_vec}{q-dimensional data vector.}

\item{m}{bandwith parameter specifying the number of Fourier frequencies
used for the estimation usually \code{floor(1+T^delta)}, where 0<delta<1.}

\item{q}{dimension of the process.}
}
\description{
\code{G.hat} Estimates the matrix G of a multivariate long memory process 
             based on an estimate of the vector of memory parameters. The assumed spectral 
             density is that of Shimotsu (2007).
}
\references{
Shimotsu, K. (2007): Gaussian semiparametric estimation of multivariate
fractionally integrated processes. Journal of Econometrics, Vol. 137, No. 2, pp. 277 - 310.
}
\keyword{internal}
