% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NS07_LM.R
\name{G_hat_nonstat}
\alias{G_hat_nonstat}
\title{Estimate G-matrix with periodogram of differenced data for NS07.}
\usage{
G_hat_nonstat(X, d, m)
}
\description{
Estimate G-matrix with periodogram of differenced data for NS07.
}
\keyword{internal}
