% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R08_LM.R
\name{H.hat}
\alias{H.hat}
\title{Real-valued element of derivative of objective function in R08.}
\usage{
H.hat(X, d, m)
}
\description{
Real-valued element of derivative of objective function in R08.
}
\keyword{internal}
