% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R08_LM.R
\name{H.hat.star}
\alias{H.hat.star}
\title{Modified complex-valued element of derivative of objective function in R08.}
\usage{
H.hat.star(X, d, m)
}
\description{
Modified complex-valued element of derivative of objective function in R08.
}
\keyword{internal}
