% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPWN.R
\name{h_LPWN}
\alias{h_LPWN}
\title{Augmented spectral density function used in LPWN estimator.}
\usage{
h_LPWN(d, theta, lambda, R_short, R_noise)
}
\description{
Augmented spectral density function used in LPWN estimator.
}
\keyword{internal}
