% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{invert}
\alias{invert}
\title{Armadillo has trouble inverting complex matrices. This is circumvented by this function.}
\usage{
invert(X)
}
\description{
Armadillo has trouble inverting complex matrices. This is circumvented by this function.
}
\keyword{internal}
