% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hou_Perron.R
\name{J.M}
\alias{J.M}
\title{Modified concentrated local Whittle likelihood. Only for internal use. cf. Hou and Perron (2014), p. 312.}
\usage{
J.M(vec, peri, m, T)
}
\description{
Modified concentrated local Whittle likelihood. Only for internal use. cf. Hou and Perron (2014), p. 312.
}
\keyword{internal}
