% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Lambda_j}
\alias{Lambda_j}
\title{Calculate matrix Lambda_j(d) as part of the spectral density}
\usage{
Lambda_j(q, n, T, d_vec)
}
\description{
Calculate matrix Lambda_j(d) as part of the spectral density
}
\keyword{internal}
