% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll_VARFIMA.R
\name{ll.VARFIMA}
\alias{ll.VARFIMA}
\title{Log-likelihood function of a VARFIMA(1,1) in final equations form.}
\usage{
ll.VARFIMA(theta, data, q, approx = 100, pre.sample = matrix(0, approx,
  q), rep = FALSE)
}
\arguments{
\item{theta}{parameter vector.}

\item{data}{data matrix with T observations of q-dimensional process.}

\item{q}{dimension of the process.}

\item{approx}{order of the AR-approximation that is supposed to be used. Default is \code{approx=100}.}

\item{pre.sample}{if likelihood is conditioned on previous observations pre.sample is an additional sample matrix.}

\item{rep}{determines whether the parameter vector is printed.}
}
\description{
\code{ll.VARFIMA} returns the value of the log-likelihood function for a given sample and parameter vector.
}
\references{
Lutkepohl, H. (2007): New introduction to multiple time series analysis. Springer.
}
