% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ll_inner}
\alias{ll_inner}
\title{calculate residuals}
\usage{
ll_inner(data, pre_sample, T, d_vec, phi, THETA, q, approx)
}
\arguments{
\item{data}{data matrix with T observations of q-dimensional process.}

\item{q}{dimension of the process.}

\item{approx}{order of the AR-approximation that is supposed to be used. Default is \code{approx=100}.}
}
\description{
calculate residuals
}
\keyword{internal}
