% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CH06_LM.R
\name{local_W_taper}
\alias{local_W_taper}
\title{Local Whittle estimator for tapered periodogram}
\usage{
local_W_taper(data, m, diff_param = 1, int = c(-0.5, 1))
}
\description{
Local Whittle estimator for tapered periodogram
}
\keyword{internal}
