% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CH03_LM.R, R/CH06_LM.R
\name{lW_wrap}
\alias{lW_wrap}
\title{Helper for local Whittle estimation.}
\usage{
lW_wrap(data, m)

lW_wrap(data, m)
}
\description{
Helper for local Whittle estimation.

Helper for local Whittle estimation.
}
\keyword{internal}
