% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WWC15_LM.R
\name{Peri_sym}
\alias{Peri_sym}
\title{Periodogram for positive and negative Fourier frequencies.}
\usage{
Peri_sym(X)
}
\description{
Periodogram for positive and negative Fourier frequencies.
}
\keyword{internal}
