% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CH03_LM.R, R/CH06_LM.R
\name{peri_taper}
\alias{peri_taper}
\title{Tapered periodogram. Only for internal use.}
\usage{
peri_taper(X, diff_param = 1)

peri_taper(X, diff_param = 1)
}
\description{
Tapered periodogram. Only for internal use.

Tapered periodogram. Only for internal use.
}
\keyword{internal}
