% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R08_LM.R
\name{Phi_j}
\alias{Phi_j}
\title{Modified version of Lambda in spectral density in R08}
\usage{
Phi_j(X, dvec)
}
\description{
Modified version of Lambda in spectral density in R08
}
\keyword{internal}
