% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_W.R
\name{PHI_lw}
\alias{PHI_lw}
\title{Scaling factor in the asymptotic variance. Only for internal use. Cf. Velasco (1999).}
\usage{
PHI_lw(ht, T, p)
}
\description{
Scaling factor in the asymptotic variance. Only for internal use. Cf. Velasco (1999).
}
\keyword{internal}
