% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPWN.R
\name{poly_LPWN}
\alias{poly_LPWN}
\title{Polynomial function used for approximation of short memory terms and noise dynamics.}
\usage{
poly_LPWN(lambda, theta_p)
}
\description{
Polynomial function used for approximation of short memory terms and noise dynamics.
}
\keyword{internal}
