% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preWhite.R
\name{pre.White}
\alias{pre.White}
\title{Pre-whitening for application of semiparametric long memory estimator.}
\usage{
pre.White(theta, data, q, approx = 100)
}
\arguments{
\item{theta}{estimated parameter vector.}

\item{data}{data matrix with T observations of q-dimensional process.}

\item{q}{dimension of the process.}

\item{approx}{order of the AR-approximation that is supposed to be used. Default is \code{approx=100}.}
}
\description{
Given a parameter vector \code{theta} obtained using \code{VARFIMA_est}, 
\code{pre.White} returns the pre-whitened sample.
}
\details{
add details here.
}
\references{
Sibbertsen, P., Leschinski, C. H., Holzhausen, M., (2015): A Multivariate Test 
             Against Spurious Long Memory. Hannover Economic Paper.
}
