% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELW_est.R
\name{R.elw}
\alias{R.elw}
\title{concentrated likelihood function for ELW estimator}
\usage{
R.elw(d, data, m)
}
\description{
concentrated likelihood function for ELW estimator
}
\keyword{internal}
