% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELW_est.R
\name{R.elw.weighted}
\alias{R.elw.weighted}
\title{concentrated likelihood function for ELW estimator - weighted version}
\usage{
R.elw.weighted(d, data, m)
}
\description{
concentrated likelihood function for ELW estimator - weighted version
}
\keyword{internal}
