% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPWN.R
\name{R.LPWN}
\alias{R.LPWN}
\title{Objective function of LPWN estimator.}
\usage{
R.LPWN(param.vec, data, m, R_short = R_short, R_noise = R_noise)
}
\description{
Objective function of LPWN estimator.
}
\keyword{internal}
