% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_W.R
\name{R.lw.hc}
\alias{R.lw.hc}
\title{Concentrated local Whittle likelihood of differenced and tapered estimator of 
Hurvich and Chen (2000). Only for internal use.}
\usage{
R.lw.hc(d, peri, m, T)
}
\description{
Concentrated local Whittle likelihood of differenced and tapered estimator of 
Hurvich and Chen (2000). Only for internal use.
}
\keyword{internal}
