% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_W.R
\name{R.lw.tapered}
\alias{R.lw.tapered}
\title{Concentrated local Whittle likelihood for tapered estimate. Only for internal use. Cf. Velasco (1999).}
\usage{
R.lw.tapered(d, peri, m, p, T)
}
\description{
Concentrated local Whittle likelihood for tapered estimate. Only for internal use. Cf. Velasco (1999).
}
\keyword{internal}
