% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{R_d_multi_GSE_coint}
\alias{R_d_multi_GSE_coint}
\title{Profiled Likelihood for GSE under Cointegration}
\usage{
R_d_multi_GSE_coint(theta_vec, PERI, m, l, T, q, elements)
}
\description{
Profiled Likelihood for GSE under Cointegration
}
\keyword{internal}
