% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CH06_LM.R
\name{R_lw_ch}
\alias{R_lw_ch}
\title{Local Whittle estimator for tapered periodogram - function to optimize}
\usage{
R_lw_ch(d, data, m, diff_param = 1)
}
\description{
Local Whittle estimator for tapered periodogram - function to optimize
}
\keyword{internal}
