% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R08_LM.R
\name{s.delta}
\alias{s.delta}
\title{Derivative of multivariate local Whittle objective function in R08.}
\usage{
s.delta(G, H)
}
\description{
Derivative of multivariate local Whittle objective function in R08.
}
\keyword{internal}
