% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cointegrating_rank_estimation.R
\name{Tab}
\alias{Tab}
\title{Test for equality of two elements d_a and d_b of estimated d vector. 
This function should not be called directly. It is called as a helper by T.rho.}
\usage{
Tab(d.hat, G.est, m1, a, b, h_n)
}
\description{
Test for equality of two elements d_a and d_b of estimated d vector. 
This function should not be called directly. It is called as a helper by T.rho.
}
\keyword{internal}
