% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MV04_LM.R
\name{V_M}
\alias{V_M}
\title{Variance for test statistic MV04.}
\usage{
V_M(x, M, m, residual, type = c("no", "const", "type"))
}
\description{
Variance for test statistic MV04.
}
\keyword{internal}
