% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{W_multi}
\alias{W_multi}
\title{Helper function for MLWS test for multivariate spurious long memory.}
\usage{
W_multi(X, d_vec, m, epsilon, eta)
}
\arguments{
\item{X}{data matrix}

\item{d_vec}{estimated vector of memory parameters.}

\item{m}{bandwith parameter specifying the number of Fourier frequencies used 
for the estimation usually \code{floor(1+T^delta)}, where 0.5<delta<0.8 for consistency.}

\item{epsilon}{trimming parameter \code{epsilon=0.05} by default. Determines
minimum number of Fourier frequencies used in test statistic. For T>500 it is recommended 
to use \code{epsilon=0.02}. Confer Sibbertsen, Leschinski, Holzhausen (2015) for further details.}

\item{eta}{weight vector.}
}
\description{
Multivariate local Whittle Score type test for the null hypothesis of true long
             memory against the alternative of spurious long memory suggested by Sibbertsen, 
             Leschinski and Holzhausen (2015).
}
\details{
add details here
}
\references{
Sibbertsen, P., Leschinski, C. H., Holzhausen, M., (2015): A Multivariate Test 
             Against Spurious Long Memory. Hannover Economic Paper.
}
