% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELW_est.R
\name{wd.elw}
\alias{wd.elw}
\title{Concentrated local Whittle likelihood. Only for internal use. cf. Shimotsu and Phillips (2005), p. ???.}
\usage{
wd.elw(d)
}
\description{
Concentrated local Whittle likelihood. Only for internal use. cf. Shimotsu and Phillips (2005), p. ???.
}
\keyword{internal}
