% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdiff.R
\name{fdiff}
\alias{fdiff}
\title{Fast fractional differencing procedure of Jensen and Nielsen (2014).}
\usage{
fdiff(x, d)
}
\arguments{
\item{x}{series to be differenced}

\item{d}{memory parameter indicating order of the fractional difference.}
}
\description{
Takes the $d$th difference of the series.
}
\details{
add details here
}
\examples{
acf(fdiff(x=rnorm(500), d=0.4))
}
\references{
Jensen, A. N. and Nielsen, M. O. (2014):
A fast fractional difference algorithm, Journal of Time Series Analysis 35(5), pp. 428-436.
}

